# Make abs and abs_imag resistant to under/overflow