Currently `abs` and `abs_imag` effectively use `abs2` under the hood. For very small numbers, this causes underflow, and for very large numbers, this causes overflow. In base. `abs(z::Complex)` uses `hypot`, which takes care to avoid under/overflow. This PR uses `hypot` in `abs` and `abs_imag`.

Before Julia 1.9, `hypot` for more than 3 args was unbearably slow (https://github.com/JuliaLang/julia/issues/44336), so `abs` is only changed for 1.9 and later.

There is a performance cost. On v1.9-alpha1:
```julia
julia> using Quaternions, BenchmarkTools

julia> q = randn(QuaternionF64);

julia> q = quat(1.3, 5.6, -3.3, 0.9);

julia> q2 = quat(fill(1e-300, 4)...);
```

On main
```julia
julia> @btime abs($q);
  1.927 ns (0 allocations: 0 bytes)

julia> @btime $(Quaternions.abs_imag)($q);
  1.922 ns (0 allocations: 0 bytes)

julia> @btime abs($q2);
  1.922 ns (0 allocations: 0 bytes)

julia> @btime $(Quaternions.abs_imag)($q2);
  1.922 ns (0 allocations: 0 bytes)
```

this PR
```julia
julia> @btime abs($q);
  8.471 ns (0 allocations: 0 bytes)

julia> @btime $(Quaternions.abs_imag)($q);
  25.249 ns (0 allocations: 0 bytes)

julia> @btime abs($q2);
  8.859 ns (0 allocations: 0 bytes)

julia> @btime $(Quaternions.abs_imag)($q2);
  25.240 ns (0 allocations: 0 bytes)
```

I have no idea why after this PR `abs_imag` is more expensive than `abs`, when it does more operations.

```julia
julia> @code_typed Quaternions.abs(q)
CodeInfo(
1 ─ %1 = Base.getfield(q, :s)::Float64
│   %2 = Base.getfield(q, :v1)::Float64
│   %3 = Base.getfield(q, :v2)::Float64
│   %4 = Base.getfield(q, :v3)::Float64
│   %5 = Core.tuple(%1, %2, %3, %4)::NTuple{4, Float64}
│   %6 = invoke Base.Math._hypot(%5::NTuple{4, Float64})::Float64
└──      return %6
) => Float64

julia> @code_typed Quaternions.abs_imag(q)
CodeInfo(
1 ─ %1 = Base.getfield(q, :v1)::Float64
│   %2 = Base.getfield(q, :v2)::Float64
│   %3 = Base.getfield(q, :v3)::Float64
│   %4 = Core.tuple(%1, %2, %3)::Tuple{Float64, Float64, Float64}
│   %5 = invoke Base.Math._hypot(%4::Tuple{Float64, Float64, Float64})::Float64
└──      return %5
) => Float64
```