@stevengj 2022-12-15
I would argue that you should prioritize correctness over performance.

It's better if people using older versions should still get correct results, even if it is slower (`abs` is not often performance-critical anyway?), than that users of different Julia versions should get completely different answers.