@stevengj 2022-12-15
If the performance is a real worry, you can always duplicate the Base code in earlier versions:
```jl
if VERSION < v"1.9"
    function _hypot(x::NTuple{N,<:Number}) where {N}
        maxabs = maximum(abs, x)
        if isnan(maxabs) && any(isinf, x)
            return typeof(maxabs)(Inf)
        elseif (iszero(maxabs) || isinf(maxabs))
            return maxabs
        else
            return maxabs * sqrt(sum(y -> abs2(y / maxabs), x))
        end
     end
    Base.abs(q::Quaternion) = hypot((real(q), imag_part(q)...))
end
```