@sethaxen 2022-12-15
> I would argue that you should prioritize correctness over performance.

Agreed!

> `abs` is not often performance-critical anyway?

One of the main uses of this package is unit quaternions for rotations, which are normalized using `abs`, so this is an operation we'd like to keep performant if possible.

> If the performance is a real worry, you can always backport the Base code to earlier versions:

Good idea! I'll adapt.

@stevengj can you think of a reason why this would be the case? :point_down:  

> I have no idea why after this PR `abs_imag` is more expensive than `abs`, when it does fewer operations.