@stevengj 2022-12-16
You might need to suppress compiler inlining and constant-propagation with
```
@btime abs($(Ref(q))[]);
```
as explained [in the BenchmarkTools manual](https://juliaci.github.io/BenchmarkTools.jl/stable/manual/#Understanding-compiler-optimizations).