@sethaxen 2022-12-16
> You might need to suppress compiler inlining and constant propagation with
> 
> ```
> @btime abs($(Ref(q))[]);
> ```
> 
> as explained [in the BenchmarkTools manual](https://juliaci.github.io/BenchmarkTools.jl/stable/manual/#Understanding-compiler-optimizations).

Doesn't seem to make a difference:

```julia
julia> using Quaternions, BenchmarkTools

julia> q = randn(QuaternionF64)
QuaternionF64(-0.06595820195841512, 0.2980891652137365, 0.1349996383131589, -0.14878742850114615)

julia> @btime abs($(Ref(q))[]);
  8.439 ns (0 allocations: 0 bytes)

julia> @btime $(Quaternions.abs_imag)($(Ref(q))[]);
  25.225 ns (0 allocations: 0 bytes)
```