@stevengj 2022-12-16
If you can reproduce it just by `hypot(3.,4.,5.,6.)` vs `hypot(3.,4.,5.)`, it might be worth filing an issue.