@mikmoore 2023-01-03
If performance is a concern, we could do things like compute `abs2` and use that if it's `floatmin/eps < abs2(q) < floatmax`, only falling back to `hypot` when this fails (but that check also adds a cost). But it'd be nicer if we didn't have this special casing here and simply could call `hypot` always.

I went wild trying to optimize `Base.Math._hypot` on `IEEEFloat` inputs. I pulled out all the stops with bit-twiddling and this is the best I came up with:
```julia
# __maxabs(x::NTuple{N,T}) where {N,T<:Base.IEEEFloat} = reinterpret(T, maximum(z -> reinterpret(Signed, abs(z)), x)) # specialization of maximum(abs,x)
__maxabs_pos(x::NTuple{N,T}) where {N,T<:Base.IEEEFloat} = reinterpret(T, maximum(z -> reinterpret(Signed, z), x)) # specialization of maximum(x) for positive-signed x
__prevpow2(x::T) where T<:Base.IEEEFloat = reinterpret(T, reinterpret(Unsigned, x) & ~Base.significand_mask(T)) # truncate to a power of 2 # DOES NOT WORK FOR SUBNORMALS
__invpow2(x::T) where T<:Base.IEEEFloat = reinterpret(T, reinterpret(Unsigned, __prevpow2(floatmax(T))) - reinterpret(Unsigned, x)) # invert a power of 2 # DOES NOT WORK FOR NEGATIVES OR SUBNORMALS

function myhypot(x::NTuple{N,T}) where {N,T<:Base.IEEEFloat}
	x = map(abs, x)
	any(==(convert(T, Inf)), x) && return convert(T, Inf)
	maxabs = __maxabs_pos(x) # == maximum(abs, x)
	m = __prevpow2(__maxabs_pos((floatmin(T), maxabs))) # something roughly on the size of m, or floatmin
	mi = __invpow2(m) # this function is only slightly faster than inv
	return m * sqrt(sum(y -> abs2(y * mi), x))
end
```
I think I got all the zero/Inf/NaN behavior correct but I haven't run a full test suite. Compared to the version provided above, this brought the runtime down from 8.2->5.7ns for 3-tuples, 8.4->6.4ns for 4-tuples and 41.3->15.9ns for 16-tuples. Meanwhile, `sqrt(sum(abs2,x))` is more like 3.0ns for 3/4-tuples and 5.8ns for 16-tuples. This may not be closing the gap as much as we'd hope, but it's maybe the fastest we can do this robustly.

Note that along the way I found that I could do better with `Base.isinf(x::IEEEFloat) = abs(x)==Inf` (in the above I drop the `abs` by doing it beforehand). That specialization should probably be added.