@sethaxen 2023-01-07
Thanks @mikmoore for looking into this! I agree it is much better if this is solved upstream in `hypot`. I opened an issue in https://github.com/JuliaLang/julia/issues/47917 and so far no one seems to have found the culprit.

`abs_imag` being slow only impacts the extensions of the analytic functions. The ones most likely to be called are `exp` and `log`, which are ~3x slower with this PR.

I suggest either we merge this PR as a whole, asserting that correctness is more important than speed (and hoping the issue is eventually resolved upstream) or we at least merge the `abs` part and perhaps save `abs_imag` for later. Thoughts?