@mikmoore 2023-01-07
The simplest thing (and my thought) is to use `hypot` and wait for the upstream fixes to resolve the performance. But I have never used analytic functions of quaternions so I'm not suffering the performance penalty here. My anticipation is that the majority of users are also in that situation. Of the ones that do use them, most probably haven't and won't notice the performance difference as I don't think these often end up in hot loops.

For what it's worth, `inv` and `/` have the same over/underflow hazard and we've yet to fix them or even open an issue. Those are more likely to be encountered than issues with analytic functions, I'd venture.

I opened JuliaLang/julia#48130, which should hopefully improve `hypot` performance a little more yet. It appears to be about 1ns slower than what I posted above but is marginally less prone to overflow and much easier to read/maintain. In that PR, I failed to reproduce JuliaLang/julia#47917 so it seems that someone has already fixed it for v1.10.